refactor: fold OAuth config into the provider registry
providers/oauth-config.ts was a second provider table: a CONFIG_FACTORIES
map keyed by provider string, parallel to the real registry. Its gdrive
entry moves onto gdriveProvider as an oauthConfig() method, and
getOAuthConfig() now reads straight off the registry.

The file is deleted rather than left as a delegating shim. getOAuthConfig
depends on registerProvider() having run, which happens as a side effect
of importing providers/index.ts; a shim in the old location would have
left that dependency invisible, and an empty registry would surface as a
plausible-looking "400 Unknown provider" on every OAuth connect rather
than an obvious failure. Exporting from index.ts makes it structural.

Behavior is unchanged at all three call sites (connect-provider,
generate-oauth-state, get-provider-access-token), which only change their
import specifier. Both rejection paths still throw, which those callers
rely on as their validation gate for an arbitrary request-body provider
string: an unregistered id throws from getProvider, and a registered
provider with no oauthConfig (osf) throws explicitly. OSF continues to be
rejected by absence rather than by special-casing.

Also removes the now-dead OAuthEndpointConfig interface and the unused
`oauth?` field on StorageProvider, superseded by oauthConfig().

Full suite green: 40 suites / 257 tests.

Note: across ten runs of the suite, two runs each had one unrelated
emulator-backed suite fail (metadata-derived-upload, pending-recovery,
upload-queue -- a different one each time, all passing in isolation, none
touching OAuth config). This matches the emulator contention flake that
.github/workflows/node.js.yml already documents and caps maxWorkers for.
